+++ examples/resources/cloudflare_split_tunnel/resource.tf
@@ -19,7 +19,7 @@ resource "cloudflare_split_tunnel" "example_split_tunnel_include" {
 }
 
 # Create a device policy
-resource "cloudflare_device_policy" "developer_warp_policy" {
+resource "cloudflare_device_settings_policy" "developer_warp_policy" {
   account_id    = "f037e56e89293a057740de681ac9abbe"
   name          = "Developers"
   precedence    = 10
@@ -28,9 +28,9 @@ resource "cloudflare_device_policy" "developer_warp_policy" {
 }
 
 # Excluding *.example.com from WARP routes for a particular device policy
-resource "cloudflare_split_tunnel" "example_device_policy_split_tunnel_exclude" {
+resource "cloudflare_split_tunnel" "example_device_settings_policy_split_tunnel_exclude" {
   account_id = "f037e56e89293a057740de681ac9abbe"
-  policy_id  = cloudflare_device_policy.developer_warp_policy.id
+  policy_id  = cloudflare_device_settings_policy.developer_warp_policy.id
   mode       = "exclude"
   tunnels {
     host        = "*.example.com"