+++ internal/framework/service/rulesets/resource_test.go
@@ -7,12 +7,16 @@ import (
 	"regexp"
 	"testing"
 
+	"github.com/cloudflare/cloudflare-go"
 	"github.com/cloudflare/terraform-provider-cloudflare/internal/acctest"
 	"github.com/cloudflare/terraform-provider-cloudflare/internal/utils"
-	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 )
+
+func TestMain(m *testing.M) {
+	resource.TestMain(m)
+}
 
 func init() {
 	resource.AddTestSweepers("cloudflare_ruleset", &resource.Sweeper{
@@ -23,26 +27,30 @@ func init() {
 			zoneID := os.Getenv("CLOUDFLARE_ZONE_ID")
 
 			if err != nil {
-				return fmt.Errorf("error establishing client: %s", err)
+				return fmt.Errorf("error establishing client: %w", err)
 			}
 
 			ctx := context.Background()
-			accountRulesets, _ := client.ListAccountRulesets(ctx, accountID)
+			accountRulesets, err := client.ListRulesets(ctx, cloudflare.AccountIdentifier(accountID), cloudflare.ListRulesetsParams{})
+			if err != nil {
+				return fmt.Errorf("failed to fetch rulesets: %w", err)
+			}
+
 			for _, ruleset := range accountRulesets {
 				if ruleset.Kind != "managed" {
-					err := client.DeleteAccountRuleset(ctx, accountID, ruleset.ID)
+					err := client.DeleteRuleset(ctx, cloudflare.AccountIdentifier(accountID), ruleset.ID)
 					if err != nil {
-						tflog.Error(ctx, fmt.Sprintf("failed to delete ruleset %q: %s", ruleset.ID, err))
+						return fmt.Errorf("failed to delete ruleset %q: %w", ruleset.ID, err)
 					}
 				}
 			}
 
-			zoneRulesets, _ := client.ListZoneRulesets(ctx, zoneID)
+			zoneRulesets, _ := client.ListRulesets(ctx, cloudflare.ZoneIdentifier(zoneID), cloudflare.ListRulesetsParams{})
 			for _, ruleset := range zoneRulesets {
 				if ruleset.Kind != "managed" {
-					err := client.DeleteZoneRuleset(ctx, zoneID, ruleset.ID)
+					err := client.DeleteRuleset(ctx, cloudflare.ZoneIdentifier(zoneID), ruleset.ID)
 					if err != nil {
-						tflog.Error(ctx, fmt.Sprintf("failed to delete ruleset %q: %s", ruleset.ID, err))
+						return fmt.Errorf("failed to delete ruleset %q: %w", ruleset.ID, err)
 					}
 				}
 			}